made inventory endpoint for ReportController
it works not perfectly bc I'm not sure how to fetch emplyee id correctly so it is just 1 for now and also createon of JSON file idk, I made it somehow but definitely need changes. Also I touched a lot of other stuff bc it was not properly working with DB and I encountered too much errors, so I needed to changes classes slightly. Also I did a little changes to DB because of naming