Update: Receipt & More
I have added a new Receipt Model as well as a DTO and a Mapper for it. I have also updated the Schema with a new Receipt entity that has a foreign key to Sale_ID - this to get information on the sale. IMPORTANT to note, the OneToOne relationship between Receipt and Sale is currently unidirectional: Receipt -> Sale. But if we ever need it to go both ways this is few lines of code.